This is an Android Mobile App “Beer Liker”.
App was built on Eclipse Juno.

The idea of this App is:
The app should display a beer from the beers API:
http://www.brewerydb.com/developers/docs
The User is able to swipe left or right images of beers, and also to
“like” or “dislike” the beers.
Alternatively the User has the option to use buttons to achieve the same
result.
After the user has liked 10 beers, they are being shown as a list of the
beers they have just liked.
The presentation of this list has been done via an custom dialog view.
Requirements
- The app should be aimed at Android 5.x.
- The app must be built using Gradle.
- Picasso should be used to load and display images.
- Try to implement the Retrofit library for network calls.

For now, when the user swipes left or right the images to “like” or
“dislike” the beer, or presses “like” or “dislike” buttons – there is no
recording are being made in order to record their “liked” beers (once I
have completed this function I will update my Project).
Although I am still using 10 pictures of beers for the “ViewPager” and
for the custom “Dialog” View.
The whole project up to date is available to anyone who wants to study
it and probably to learn something new from it.
Please don’t judge too hard on my design, it was just me playing with
it.
This Project is not for commercial publications, and developer reserves
Project’s Copy Rights (contact me for more information).
Anyone who wishes to carry on with my work, to complete Project’s
requirements, please feel free as long as you keen to share your final
results with me, so I might learn something new myself … Thanks in
advance.
I hope you guys will enjoy it.